Changing bigint to int for entities hash

+++ kifi-backend/modules/common/conf/evolutions/shoebox/381.sql
@@ -2,7 +2,7 @@
 
 # --- !Ups
 
-ALTER TABLE bookmark ADD COLUMN `entities_hash` bigint(20) DEFAULT NULL; -- for now!
+ALTER TABLE bookmark ADD COLUMN `entities_hash` int(10) DEFAULT NULL; -- for now!
 ALTER TABLE bookmark ADD INDEX bookmark_i_entities_hash (entities_hash);
 
 insert into evolutions (name, description) values('381.sql', 'add unique column bookmark.entities_hash');
